requirements.txt
Appium-Python-Client==1.2.0
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.9
docutils==0.17.1
kitchen==1.2.6
pip==21.2.4
robotframework==4.1
robotframework-appiumlibrary==1.6.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==5.1.3
selenium==3.141.0
setuptools==57.5.0
six==1.16.0
wheel==0.37.0

